+++ classes/SmartResizer.class.php
@@ -1,14 +1,13 @@
 <?php
 /**
  * Resize and cache images according to "img src" tags.
- * Based on the SmartResizer plugin for Joomla.
+ * Ignore images that are already part of an existing link.
  *
- * @see https://extensions.joomla.org/extension/smartresizer/
- * @author     Lee Garner <lee@leegarner.com>
- * @copyright  Copyright (c) 2017 Lee Garner <lee@leegarner.com>
- * @package    lglib
- * @version    1.0.5
- * @license    http://opensource.org/licenses/gpl-2.0.php
+ * @author      Lee Garner <lee@leegarner.com>
+ * @copyright   Copyright (c) 2017-2019 Lee Garner <lee@leegarner.com>
+ * @package     lglib
+ * @version     v1.0.9
+ * @license     http://opensource.org/licenses/gpl-2.0.php
  *              GNU Public License v2 or later
  * @filesource
  */
@@ -17,7 +16,8 @@ namespace LGLib;
 /**
  * SmartResizer class.
  */
-class SmartResizer {
+class SmartResizer
+{
 
     /**
      * Resize images found in a template.
@@ -49,36 +49,29 @@ class SmartResizer {
     {
         global $_CONF;
 
-        $runword='';    // Holdover from Joomla. Trigger to skip processing
-        $regex_img = "|<[\s\v]*img[\s\v]([^>]*".$runword."[^>]*)>|Ui";
-        preg_match_all( $regex_img, $origtxt, $matches_img);
-        $count_img = count($matches_img[0]);
-
-        for ($i = 0; $i < $count_img; $i++) {
-            // Skip processing if tagged not to resize
-            if (strpos($matches_img[0][$i], 'nosmartresize'))
-                continue;
-
-            // Skip processing if an invalid image is found
-            if (!@$matches_img[1][$i])
-                continue;
-
-            // Initialize vars. $inline_params contains the entire img tag
-            $image_width = 0;
-            $image_height = 0;
-            $inline_params = $matches_img[1][$i];
-            $src = array();
-
-            // Get the "src=" part of the tag.
-            preg_match( "#src=\"(.*?)\"#si", $inline_params, $src );
-            if (isset($src[1])) {
-                $src = trim($src[1]);
-            } else {
-                // no img src value found
+        $dom= new \DOMDocument();
+        $dom->loadHTML($origtxt);
+        $xpath = new \DOMXPath($dom);
+        $images = $xpath->query("//img");
+        foreach ($images as $img) {
+            // save the entire tag <img src="..." class=... />
+            $tag = $img->ownerDocument->saveXML($img);
+            if (strpos($tag, 'noresize') > 0) {
                 continue;
             }
 
+            $is_linked = false;
+            $p = $img;
+            while ($p->parentNode !== NULL) {
+                if ($p->tagName == 'a') {
+                    $is_linked = true;
+                    break;
+                }
+                $p = $p->parentNode;
+            }
+
             // Split up the src, check if it's a relative or remote URL
+            $src = $img->getAttribute('src');
             $url_parts = parse_url($src);
             if (isset($url_parts['host'])) {
                 continue;  // don't handle remote images
@@ -87,21 +80,6 @@ class SmartResizer {
             // Split up the path to extract the filename and extension
             $fparts = pathinfo($url_parts['path']);
             $extension = $fparts['extension'];
-            /*switch ($extension) {
-            case 'jpg':
-            case 'jpeg':
-                $mime = 'image/jpeg';
-                break;
-            case 'png':
-                $mime = 'image/png';
-                break;
-            case 'bmp':
-                $mime = 'image/bmp';
-                break;
-            case 'gif':
-                $mime = 'image/gif';
-                break;
-            }*/
 
             // Create the thumbnail path in "/thumbs" under the original dir
             $thumb_path = str_replace('/', DIRECTORY_SEPARATOR,
@@ -115,12 +93,41 @@ class SmartResizer {
             // Get the height and width parameters, if supplied.
             // If one is missing, reDim() will resize based on the supplied one.
             // If both are missing, then continue since there's no resizing to do.
-            $d_width = self::getDimFromTag('width', $inline_params);	
-            $d_height= self::getDimFromTag('height', $inline_params);	
+            $d_width = 0;
+            $d_height = 0;
+            if ($img->hasAttribute('style')) {
+                $attribs = explode(';', $img->getAttribute('style'));
+                foreach ($attribs as $attr) {
+                    $parts = explode(':', $attr);
+                    if ($parts[0] == 'width') {
+                        $d_width = (int)$parts[1];
+                    } elseif ($parts[0] == 'height') {
+                        $d_height = (int)$parts[1];
+                    }
+                }
+            }
+            // Now look for old-style tags if the dimenstions aren't already included
+            if ($d_width == 0) {
+                $val = $img->getAttribute('width');
+                if (!empty($val)) {
+                    $d_width = $val;
+                }
+            }
+            if ($d_height== 0) {
+                $val = $img->getAttribute('height');
+                if (!empty($val)) {
+                    $d_height = $val;
+                }
+            }
+            // If there isn't at least one, then give up, no resizing to do
             if (empty($d_width) && empty($d_height)) {
                 continue;
             }
 
+            // Determine the new width and height that will fit within the specified
+            // tags while preservint the aspect ratio.
+            // Have to do this here since Image::ReSize() won't be called if the
+            // thumbnail exists, but we still need the correct image dimensions.
             $A = Image::reDim($src_path, $d_width, $d_height);
             if ($A === false) continue;
             $s_width = $A['s_width'];
@@ -129,8 +136,6 @@ class SmartResizer {
             $d_height = $A['d_height'];
             $mime = $A['mime'];
 
-            //list($s_width,$s_height, $d_width, $d_height) =
-            //        self::reDim($src_path, $d_width, $d_height);
             if ($s_width < $d_width && $s_height < $d_height) {
                 // Don't scale the image up, just use the original.
                 continue;
@@ -141,60 +146,44 @@ class SmartResizer {
                      $_CONF['path_html'] . $thumb_url);
             if (!file_exists($thumb_path)) {
                 // Thumb doesn't already exist, create it
-                if (function_exists(_img_resizeImage)) {
-                    $result = _img_resizeImage($src_path, $thumb_path,
-                            $s_height, $s_width,
-                            $d_height, $d_width, $mime);
-                }
+                $res = Image::ReSize($src_path, $thumb_path, $d_width, $d_height);
             }
+
             if (!file_exists($thumb_path)) {
                 // Something went wrong if the thumb file still doesn't exist.
                 continue;
             }
 
-            // Get the alt tag to use as a title
-    		preg_match("#alt=\"(.*?)\"#si", $inline_params, $title);
-	    	if (isset($title[1])) {
-                $title = ' title="' . trim($title[1]) . '" ';
-            } else {
-                $title = '';
-            }
+            // Set the new image source to the thumbnail URL.
+            // This is the only thing changed in the img tag.
+            $img->setAttribute('src', $thumb_url);
 
-    		$text = str_replace($src, $thumb_url, $matches_img[0][$i]);
-	        $text = '<a href="' . $src .
-                '" data-uk-lightbox="{group:\'article\'}" " width="' .
-                $s_width . '" height="' . $s_height . '" ' . $title .
-                '>'.$text.'</a>';
-	        $origtxt = str_replace($matches_img[0][$i], $text, $origtxt);
-        }
-    }
+            // Now, create a link to the original image, but only if the image
+            // isn't already part of a link tag.
+            if (!$is_linked) {
+                // Create a "a" element for the link
+                $a = $dom->createElement('a');
 
+                // Get the alt tag to use as a title
+                $title = $img->getAttribute('alt');
+                if (!empty($title)) {
+                    $a->setAttribute('title', $title);
+                }
 
-    /**
-     * Get the dimension parameters from the image source tag.
-     * First check for style tag, then look for width/height tags.
-     *
-     * @param   string  $type   "width" or "height"
-     * @param   string  $content    Image tag content
-     * @return  string      Size from tag, or empty string if not found
-     */
-    private static function getDimFromTag($type, $content)
-    {
-        $size = array();
-        preg_match("#[\s\;\"]{$type}:(.*?)px*[\s\;\"]#si", $content, $size);
-        if (isset($size[1])) {
-            $retval = trim($size[1]);
-        } else {
-            // style="width:..." tag not found, look for simple "width=xx"
-    		preg_match("#{$type}=\"(.*?)\"#si", $content, $size);
-    		if (isset($size[1])) {
-                $retval = trim($size[1]);
-            } else {
-                $retval = '';
+                $a->setAttribute('href', $src);
+                $a->setAttribute('data-uk-lightbox', "{group:'article'}");
+                $a->setAttribute('style', 'width:' . $A['s_width'] . 'px;height:' . $A['s_height'] . 'px');
+                $a->appendChild($img->cloneNode());
+
+                // replace img with the wrapper that is holding the <img>
+                $img->parentNode->replaceChild($a, $img);
             }
         }
-        return $retval;
+
+        // Finally, save the new document into the original text
+        $origtxt = $dom->saveHTML();
     }
+
 }
 
 ?>